# backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CrmApi | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5194/"
    }
  ],
  "paths": {
...
    "/api/companies/{id}": {
      "get": {
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lifecycleStage",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/deals/{dealId}": {
      "get": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pipeline"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DealDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/companies": {
      "get": {
        "tags": [
          "PublicCompanies"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicCompanyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicCompanyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicCompanyDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{id}": {
      "get": {
        "tags": [
          "PublicCompanies"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCompanyDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCompanyDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCompanyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contacts": {
      "get": {
        "tags": [
          "PublicContacts"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicContactDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicContactDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicContactDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contacts/{id}": {
      "get": {
        "tags": [
          "PublicContacts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicContactDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicContactDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicContactDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deals": {
      "get": {
        "tags": [
          "PublicDeals"
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicDealDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicDealDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PublicDealDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/deals/{id}": {
      "get": {
        "tags": [
          "PublicDeals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDealDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDealDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicDealDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityDto": {
        "required": [
          "id",
          "type",
          "body",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "body": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CompanyDto": {
        "required": [
          "id",
          "name",
          "domain",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": [
              "null",
              "string"
            ]
          },
          "customFields": {
            "type": "object"
          }
        }
      },
      "ContactDto": {
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "companyName",
          "lifecycleStage",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastName": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "companyName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lifecycleStage": {
            "type": "string"
          },
          "customFields": {
            "type": "object"
          }
        }
      },
...
      "CreateContactRequest": {
        "required": [
          "firstName",
          "lastName",
          "email",
          "companyName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "companyName": {
            "type": [
              "null",
              "string"
            ]
          },
          "customFields": {
            "type": [
              "null",
              "object"
            ]
          }
        }
      },
...
      "DealDetailDto": {
        "required": [
          "id",
          "title",
          "stageName",
          "amountCents",
          "currency",
          "forecastCategory",
          "aiScore",
          "aiScoreRationale",
          "aiSummary",
          "aiSummarizedAt",
          "activitiesSinceSummary",
          "contactNames",
          "activities",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "stageName": {
            "type": "string"
          },
          "amountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "forecastCategory": {
            "type": "string"
          },
          "aiScore": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "aiScoreRationale": {
            "type": [
              "null",
              "string"
            ]
          },
          "aiSummary": {
            "type": [
              "null",
              "string"
            ]
          },
          "aiSummarizedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "activitiesSinceSummary": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "contactNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "activities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityDto"
            }
          },
          "customFields": {
            "type": "object"
          }
        }
      },
...
      "PublicCompanyDto": {
        "required": [
          "id",
          "name",
          "domain",
          "createdAt",
          "updatedAt",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "domain": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "customFields": {
            "type": "object"
          }
        }
      },
      "PublicContactDto": {
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "phone",
          "companyId",
          "lifecycleStage",
          "createdAt",
          "updatedAt",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "firstName": {
            "type": [
              "null",
              "string"
            ]
          },
          "lastName": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "phone": {
            "type": [
              "null",
              "string"
            ]
          },
          "companyId": {
            "type": [
              "null",
              "string"
            ]
          },
          "lifecycleStage": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "customFields": {
            "type": "object"
          }
        }
      },
      "PublicDealDto": {
        "required": [
          "id",
          "companyId",
          "pipelineId",
          "stageId",
          "amountCents",
          "currency",
          "forecastCategory",
          "contactIds",
          "closedAt",
          "createdAt",
          "updatedAt",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": [
              "null",
              "string"
            ]
          },
          "pipelineId": {
            "type": "string"
          },
          "stageId": {
            "type": "string"
          },
          "amountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "forecastCategory": {
            "type": "string"
          },
          "contactIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "closedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "customFields": {
            "type": "object"
          }
        }
      },
...
      "UpdateCompanyRequest": {
        "required": [
          "name",
          "domain",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "domain": {
            "type": [
              "null",
              "string"
            ]
          },
          "customFields": {
            "type": [
              "null",
              "object"
            ]
          }
        }
      },
      "UpdateDealRequest": {
        "required": [
          "amountCents",
          "currency",
          "forecastCategory",
          "customFields"
        ],
        "type": "object",
        "properties": {
          "amountCents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "forecastCategory": {
            "type": "string"
          },
          "customFields": {
            "type": [
              "null",
              "object"
            ]
          }
        }
      },
...
    }
  },
...
}